Add 2 new rules: AMI_OUTDATED + VPC_SG_OPEN_TO_AUTHORIZED_PORT_ONLY + cleanup
