+++ python/ec2_desired_instance_type.py
